requirements.txt
boto3==1.16.48
python-dateutil==2.8.1
